hotfix for match tokenizer
