+++ package.json
@@ -2,7 +2,7 @@
   "author": "Spencer Kelly <spencermountain@gmail.com> (http://spencermounta.in)",
   "name": "compromise",
   "description": "modest natural language processing",
-  "version": "13.10.3",
+  "version": "13.10.4",
   "main": "./builds/compromise.js",
   "unpkg": "./builds/compromise.min.js",
   "module": "./builds/compromise.mjs",
@@ -41,8 +41,9 @@
     "coverage": "nyc -r lcov -n 'src/**/*' -n 'plugins/**/*' npm run test",
     "codecov": "npm run coverage && codecov -t 15039ad1-b495-48cd-b4a0-bcf124c9b318",
     "perf": "node ./scripts/perf/index.js",
+    "perf:build": "TESTENV=prod node ./scripts/perf/index.js",
     "perf:versions": "node ./scripts/perf/versions.js",
-    "perf:build": "node ./scripts/perf/build-speed.js",
+    "flame": "clinic flame -- node ./scripts/perf/flame",
     "lint": "eslint ./src/ && eslint ./plugins/**/src/",
     "watch": "amble ./scratch.js",
     "build:all": "node ./scripts/build/build-all.js && npm run build --silent",
