+++ plugins/match-runner/src/fastTagger.js
@@ -1,4 +1,6 @@
 const fastTagger = function (doc, matches) {
+  doc.cache()
+  console.log(doc._cache)
   return doc
 }
 module.exports = fastTagger
