+++ scripts/perf/flame/_sotu-text.js
