+++ scripts/perf/pool/_lib.js
@@ -2,7 +2,7 @@ if (typeof process !== undefined && typeof module !== undefined) {
   let nlp
   if (process.env.TESTENV === 'prod') {
     console.warn('== production build test 🚀 ==')
-    nlp = require('../../../types')
+    nlp = require('../../../builds/compromise.min.js')
   } else {
     nlp = require('../../../src')
     // nlp.extend(require('../plugins/numbers/src'))
